test(feed): keep PostgreSQL proof in the approved composition root
